backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
tensorflow==2.15.0
Pillow==10.1.0
numpy==1.24.3
python-jose==3.3.0